[fix] : Move the logged-in imac API to the 'location' domain
