+++ src/main/java/kr/where/backend/cluster/swagger/ClusterApiDocs.java
@@ -6,26 +6,9 @@ import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import kr.where.backend.auth.authUser.AuthUser;
-import kr.where.backend.auth.authUser.AuthUserInfo;
-import kr.where.backend.cluster.dto.ResponseClusterListDTO;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
 public interface ClusterApiDocs {
-    @Operation(summary = "getCluster API", description = "클러스터별 imac 로그인한 사람 조회",
-            parameters = {
-                    @Parameter(name = "accessToken", description = "인증/인가 확인용 accessToken", in = ParameterIn.HEADER),
-            },
-            responses = {
-                @ApiResponse(responseCode = "200", description = "imac에 로그인한 멤버 조회 성공", content = @Content(schema = @Schema(implementation = ResponseClusterListDTO.class))),
-            }
-    )
-    @GetMapping("/{cluster}")
-    public ResponseEntity<ResponseClusterListDTO> getCluster(@PathVariable("cluster") final String cluster, @AuthUserInfo final AuthUser authUser);
-
     @Operation(summary = "initCluster API", description = "DB의 Cluster 데이터 모두 초기화",
             parameters = {
                     @Parameter(name = "accessToken", description = "인증/인가 확인용 accessToken", in = ParameterIn.HEADER),
@@ -36,5 +19,4 @@ public interface ClusterApiDocs {
     )
     @PostMapping("/init")
     public void init();
-
 }
