+++ src/main/java/kr/where/backend/location/LocationController.java
@@ -2,6 +2,8 @@ package kr.where.backend.location;
 
 import jakarta.validation.Valid;
 import kr.where.backend.auth.authUser.AuthUser;
+import kr.where.backend.auth.authUser.AuthUserInfo;
+import kr.where.backend.location.dto.ResponseLoggedImacListDTO;
 import kr.where.backend.location.dto.ResponseLocationDTO;
 import kr.where.backend.location.dto.UpdateCustomLocationDTO;
 import kr.where.backend.location.swagger.LocationApiDocs;
@@ -10,6 +12,8 @@ import lombok.extern.slf4j.Slf4j;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -51,4 +55,9 @@ public class LocationController implements LocationApiDocs {
 		return ResponseEntity.ok(responseLocationDto);
 	}
 
+	@GetMapping("/active/{cluster}")
+	public ResponseEntity<ResponseLoggedImacListDTO> getLoggedInIMacs(@PathVariable("cluster") final String cluster,
+																	  @AuthUserInfo final AuthUser authUser) {
+		return ResponseEntity.ok(locationService.getLoggedInIMacs(authUser, cluster));
+	}
 }
